ui(additional-lines): move picker above Analyze & Suggest, add notice
Ties the "Additional lines to prevent flow increase" picker to the
recommender analysis: it now sits directly above the Analyze &
Suggest button and shares its visibility — hidden during Analyzing…
and after results are pending, exactly like the button itself.

- New `AdditionalLinesPicker` component (extracted from OverloadPanel)
  receives `branches` / `n1Overloads` / `additionalLinesToCut` /
  `onToggle` and renders chips + a typeahead input.
- ActionFeed renders the picker only in the run-analysis branch of
  its analysis-action slot, so it disappears the moment Analyze &
  Suggest is clicked.
- OverloadPanel reverts to its pre-extraction shape (no more
  branches / additionalLinesToCut props).
- New `additional-lines-to-cut` warning notice surfaces in the
  sidebar Notices panel whenever the picker is non-empty AND no
  analysis result is pending — so the operator can see at a glance
  which extra lines the next run will treat as flow-increase
  prevention targets.
